+++ assets/appFramework/angular/generate/directive/unit/mochachai/gen-name-snake.directive.spec.ts
@@ -1,7 +1,7 @@
 /**
  * Tests for GenNamePascal directive.
  */
-import /* Synthetic Import */ chai from "chai";
+import chai from "chai";
 import { GenNamePascalDirective } from "../../src/gen-name-snake.directive";
 
 describe("GenNamePascalDirective", () => {