+++ assets/appFramework/angular/generate/module/unit/mochachai/gen-name-snake.module.spec.ts
@@ -1,7 +1,7 @@
 /**
  * Tests for GenNamePascalModule.
  */
-import /* Synthetic Import */ chai from "chai";
+import chai from "chai";
 import { GenNamePascalModule } from "../../src/gen-name-snake.module";
 
 describe("GenNamePascalModule", () => {