+++ assets/appFramework/angular/generate/pipe/unit/mochachai/gen-name-snake.pipe.spec.ts
@@ -1,7 +1,7 @@
 /**
  * Tests for GenNamePascal pipe.
  */
-import /* Synthetic Import */ chai from "chai";
+import chai from "chai";
 import { GenNamePascalPipe } from "../../src/gen-name-snake.pipe";
 
 describe("GenNamePascalPipe", () => {