+++ assets/appFramework/aurelia/generate/value-converter/unit/mochachai/gen-name-snake.spec.ts
@@ -1,7 +1,7 @@
 /**
  * Tests for GenNamePascalValueConverter.
  */
-import /* Synthetic Import */ chai from "chai";
+import chai from "chai";
 import { GenNamePascalValueConverter } from "../../src/gen-name-snake";
 
 describe("GenNamePascalValueConverter", () => {