+++ src/pipelineSteps/moduleType/amd.ts
@@ -17,7 +17,6 @@ export class Amd extends PipelineStepBase {
 
     public async initialise(logger: ILogger, fileSystem: IFileSystem, uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables, mainCondition: boolean): Promise<number> {
         if (mainCondition) {
-            engineVariables.syntheticImport = "* as ";
             engineVariables.moduleId = "module.id.toString()";
         }
         return 0;