+++ test/unit/src/pipelineSteps/moduleType/amd.spec.ts
@@ -62,18 +62,18 @@ describe("Amd", () => {
 
     describe("initialise", () => {
         it("can be called with not matching condition", async () => {
-            engineVariablesStub.syntheticImport = "foo";
+            engineVariablesStub.moduleId = "foo";
             const obj = new Amd();
             const res = await obj.initialise(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, false);
             Chai.expect(res).to.be.equal(0);
-            Chai.expect(engineVariablesStub.syntheticImport).to.be.equal("foo");
+            Chai.expect(engineVariablesStub.moduleId).to.be.equal("foo");
         });
 
         it("can be called with matching condition", async () => {
             const obj = new Amd();
             const res = await obj.initialise(loggerStub, fileSystemMock, uniteConfigurationStub, engineVariablesStub, true);
             Chai.expect(res).to.be.equal(0);
-            Chai.expect(engineVariablesStub.syntheticImport).to.be.equal("* as ");
+            Chai.expect(engineVariablesStub.moduleId).to.be.equal("module.id.toString()");
         });
     });