+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Added babelify to browserify bundling
 * Added babel-loader to webpack bundling
 * Change clientPackage add command defaults to * for main if none exists in package.json
+* Replaced synthetic import generation with TypeScript esModuleInterop option
 
 ## 1.8.0